#6 add liquibase base setup
