+++ localization-finder-core/src/main/resources/liquibase/localization-finder-changelog.xml
@@ -1,0 +1,8 @@
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.2.xsd">
+
+    <include file="sprint-01/sprint-01.xml" relativeToChangelogFile="true"/>
+
+</databaseChangeLog>
