+++ localization-finder-core/src/main/resources/liquibase/sprint-01/ddl/01-add-database-schema.xml
@@ -1,0 +1,54 @@
+<databaseChangeLog
+        xmlns="http://www.liquibase.org/xml/ns/dbchangelog"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-3.2.xsd">
+
+    <changeSet id="sprint01_ddl_01-add-database-schema.xml" author="ptarkowski">
+
+        <createSequence sequenceName="country_id_seq" startValue="1" incrementBy="1"/>
+        <createSequence sequenceName="localization_id_seq" startValue="1" incrementBy="1"/>
+        <createSequence sequenceName="city_id_seq" startValue="1" incrementBy="1"/>
+
+        <createTable tableName="country">
+            <column name="id" type="bigint" valueSequenceNext="country_id_seq">
+                <constraints primaryKey="true"/>
+            </column>
+            <column name="name" type="varchar(255)">
+                <constraints nullable="false"/>
+            </column>
+        </createTable>
+
+        <createTable tableName="city">
+            <column name="id" type="bigint" valueSequenceNext="city_id_seq">
+                <constraints primaryKey="true"/>
+            </column>
+            <column name="name" type="varchar(255)">
+                <constraints nullable="false"/>
+            </column>
+            <column name="country_id" type="bigint">
+                <constraints referencedTableName="country" referencedColumnNames="id"
+                             foreignKeyName="fk_city_to_country" nullable="false"/>
+            </column>
+        </createTable>
+
+        <createTable tableName="localization">
+            <column name="id" type="bigint" valueSequenceNext="localization_id_seq">
+                <constraints primaryKey="true"/>
+            </column>
+            <column name="address" type="varchar(255)">
+                <constraints nullable="false"/>
+            </column>
+            <column name="city_id" type="bigint">
+                <constraints referencedTableName="city" referencedColumnNames="id"
+                             foreignKeyName="fk_localization_to_city" nullable="false"/>
+            </column>
+            <column name="type" type="varchar(30)">
+                <constraints nullable="false"/>
+            </column>
+            <column name="zip_code" type="varchar(30)"/>
+            <column name="latitude" type="double"/>
+            <column name="longitude" type="double"/>
+        </createTable>
+    </changeSet>
+
+</databaseChangeLog>
