+++ localization-finder-core/src/test/java/pl/tarko/software/finder/CorePreConfiguredContextTest.java
@@ -1,6 +1,10 @@
 package pl.tarko.software.finder;
 
 import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
@@ -8,5 +12,6 @@ import org.springframework.test.context.junit4.SpringRunner;
 @RunWith(SpringRunner.class)
 @ActiveProfiles("test")
 @ContextConfiguration(classes = {LocalizationCoreConfiguration.class})
+@SpringBootTest
 public abstract class CorePreConfiguredContextTest {
 }
