+++ localization-finder-core/src/test/resources/application-test.properties
@@ -1,6 +1,8 @@
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.database=hsql
 spring.datasource.driver-class-name=org.hsqldb.jdbcDriver
 spring.datasource.url=jdbc:hsqldb:mem:localization-core-test
 spring.datasource.username=sa
 spring.datasource.password=
+liquibase.change-log=classpath:/liquibase/localization-finder-changelog.xml
+liquibase.contexts=test
