+++ localization-finder-main/src/main/resources/application.properties
@@ -1,9 +1,11 @@
-spring.jpa.hibernate.ddl-auto=create-drop
+spring.jpa.hibernate.ddl-auto=validate
 spring.jpa.database=hsql
 spring.jpa.show-sql=true
 spring.datasource.driver-class-name=org.hsqldb.jdbcDriver
-spring.datasource.url=jdbc:hsqldb:mem:localization-dev
+spring.datasource.url=jdbc:hsqldb:file:db/localization-dev;shutdown=true
 spring.datasource.username=sa
 spring.datasource.password=
 googlemaps.http.client.timeout=10000
 googlemaps.geocoder.client.key=AIzaSyAKxelGlKpPruFWlJJsciSz0quyDvNZZIg
+liquibase.change-log=classpath:/liquibase/localization-finder-changelog.xml
+liquibase.contexts=dev
