+++ localization-finder-main/src/test/resources/application-test.properties
@@ -4,3 +4,5 @@ spring.datasource.driver-class-name=org.hsqldb.jdbcDriver
 spring.datasource.url=jdbc:hsqldb:mem:localization-web-test
 spring.datasource.username=sa
 spring.datasource.password=
+liquibase.change-log=classpath:/src/main/resources/liquibase/localization-finder-changelog.xml
+liquibase.contexts=test
